Rename package
A sub package snuck in there.  I don't want it right now.  I might change my mind later
