+++ src/gameComponents/DemoMap.java
@@ -1,4 +1,4 @@
-package gameComponents.Background;
+package gameComponents;
 
 import BBDGameLibrary.GameEngine.GameItem2d;
 import BBDGameLibrary.Geometry2d.BBDPolygon;
